Extract out reuseable parts of Meta Macro
